requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-countries==5.5
django-crispy-forms==1.7.2
idna==2.8
Pillow==6.1.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
stripe==2.35.1
urllib3==1.25.3
